Fixing Mesh Bug
I am fixing the mesh bug. The problem was either the on or reset pins going to the mesh. I need to test them individually and get them up and running again. Still have to integrate Dane's functions into a final class, and think on how to simplify the API packet classes. Also need to work with Dane to extend mesh functionality.